Trying setup app with preloaded scrapers on db
