+++ pcts_scrapers_api/scrapers/tasks.py
@@ -7,6 +7,7 @@ from celery import Celery
 from celery import shared_task, task
 from celery.schedules import crontab
 from celery.canvas import group, chain, chord
+from pcts_scrapers_api.celery import app as celery_app
 
 if (os.environ.get("PROJECT_ENV_EXECUTOR", default="HOST") == "DOCKER"):
     sys.path.append('/app/pcts_scraper_jobs')
@@ -15,46 +16,26 @@ else:
 
 from scraper_executor import run_scraper
 
+from scrapers.models import Scraper
+
 # from scrapers.models import ScraperExecutionGroup
 # from scrapers.models import ScraperExecution
 
 
-def task_scraper_group_wrapper(task_group_name, task_sub_prefix_name, scraper_id):
+def task_scraper_group_wrapper(task_group_name, task_sub_prefix_name):
     @task(name=task_sub_prefix_name)
-    def task_scraper_subtask(keyword, **kwargs):
-        run_scraper(scraper_id, keyword)
-
-        # Save ScraperExecutionGroup
-        # group = ScraperExecutionGroup.objects.create(
-        #     scraper=
-        # )
-
-        # scraper = models.ForeignKey(Scraper, on_delete=models.CASCADE)
-        # task_name = models.CharField("Task Name", max_length=100)
-        # start_datetime = models.DateTimeField("Start Datetime", auto_now_add=True)
-        # end_datetime = models.DateTimeField("End Datetime", null=True)
-        # status = models.IntegerField(
-        #     "Execution Status",
-        #     choices=STATUS_CHOICES,
-        #     default=1
-        # )
-
-        # ScraperExecution.objects.create(
-        #     scraper_execution_group=group,
-
-        # )
-
-
-    # scraped_pages = models.IntegerField()
-    # saved_records = models.IntegerField()
-    # droped_records = models.IntegerField()
+    def task_scraper_subtask(scraper_classname, keyword, **kwargs):
+        run_scraper(scraper_classname, keyword)
         return True
 
     @task(name=task_group_name)
-    def task_scraper_group(keywords, **kwargs):
+    def task_scraper_group(scraper_classname, keywords, **kwargs):
         task_scraper_subtasks = [
             task_scraper_subtask.subtask(
-                kwargs={"keyword": keyword},
+                kwargs={
+                    "keyword": keyword,
+                    "scraper_classname": scraper_classname
+                },
                 immutable=True
             )
             for keyword in keywords
@@ -63,3 +44,34 @@ def task_scraper_group_wrapper(task_group_name, task_sub_prefix_name, scraper_id
         return True
 
     return task_scraper_group
+
+
+# ============================= AUTO CREATE SCHEDULERS ON STARTUP
+@celery_app.on_after_finalize.connect
+def setup_periodic_scrapers(sender: Celery, **kwargs):
+    KEYWORDS = [
+        "povos e comunidades tradicionais",
+        "quilombolas",
+    ]
+    print("ADICIONANDO PERIODIC TASKS")
+
+    DEFAULT_SCRAPERS = [
+        {"id": 1, "class": "MpfScraperSpider"},
+        {"id": 2, "class": "IncraScraperSpider"},
+    ]
+
+    for scraper_config in DEFAULT_SCRAPERS:
+        scraper = Scraper.objects.get(pk=scraper_config["id"])
+
+        sender.add_periodic_task(
+            crontab(minute='0', hour='4', day_of_week='*',
+                    day_of_month='*', month_of_year='*'),
+            task_scraper_group_wrapper(
+                scraper.task_name_prefix,
+                f"{scraper.task_name_prefix}_keyword",
+            ).subtask(kwargs={
+                "keywords": KEYWORDS,
+                "scraper_classname": scraper_config["class"],
+            }),
+            name=scraper.task_name_prefix,
+        )
