+++ requirements.txt
@@ -16,6 +16,7 @@ celery==4.4.7
 django-celery-results
 django-celery-beat
 flower==0.9.5
+vine
 
 # Django API
 django
